pyproject.toml
[tool.poetry]
name = "forestreebackend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.7"
channels = {extras = ["daphne"], version = "^4.0.0"}
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^3.13.0"
pillow = "^9.4.0"
stripe = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
